refactor linear array ops helper
